refactor(resource): relocate package
