+++ core/resource/locator.go
