+++ core/resource/store_file.go
