+++ core/resource/store_mem.go
