+++ core/resource/util.go
