+++ game/controller.go
@@ -2,10 +2,10 @@ package game
 
 import (
 	"github.com/mokiat/lacking/app"
+	"github.com/mokiat/lacking/core/resource"
 	"github.com/mokiat/lacking/debug/metric"
 	"github.com/mokiat/lacking/game/graphics"
 	"github.com/mokiat/lacking/game/physics"
-	"github.com/mokiat/lacking/resource"
 	"github.com/mokiat/lacking/util/async"
 )
