+++ game/resource_registry.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"sync"
 
-	"github.com/mokiat/lacking/resource"
+	"github.com/mokiat/lacking/core/resource"
 	"github.com/mokiat/lacking/storage/chunked"
 	"github.com/mokiat/lacking/util/async"
 )
